Build tables from entities
