+++ Rocket.toml
@@ -1,11 +1,8 @@
-[default]
-template_dir = "templates/"
-
-[default.databases.rocket_example]
+[default.databases.rocket_starter]
 # Mysql
 # make sure to enable "sqlx-mysql" feature in Cargo.toml, i.e default = ["sqlx-mysql"]
 # url = "mysql://root:@localhost/rocket_example"
 
 # Postgres
 # make sure to enable "sqlx-postgres" feature in Cargo.toml, i.e default = ["sqlx-postgres"]
-url = "postgres://root:root@localhost/rocket_example"
+url = "postgres://root:root@localhost/rocket_starter"
